% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conscaper_setup.R
\name{ConScapeR_setup}
\alias{ConScapeR_setup}
\title{Provide path to the julia installation}
\usage{
ConScapeR_setup(julia_path, install_libraries = F)
}
\arguments{
\item{install_libraries}{\code{TRUE/FALSE} for whether the libraries should be installed in Julia (FALSE by default)}
}
\description{
The first time the function is ran, it is best to set the \code{install_libraries} to TRUE to install the ConScape library in Julia
}
