% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{Grid}
\alias{Grid}
\title{Wrapper for the Grid function of \code{ConScape}}
\usage{
Grid(affinities, sources, targets, costs)
}
\arguments{
\item{affinities}{\verb{[SpatRaster, matrix]}}

\item{sources}{\verb{[SpatRaster, matrix]}}

\item{targets}{\verb{[SpatRaster, matrix]}}

\item{costs}{\verb{[SpatRaster, matrix, character]}}
}
\description{
Creates a Grid from affinities, sources, targets and costs.
Affinities, sources and targets can be provided either as SpatRaster from the terra library or as matrix.
The costs can be provided as SpatRaster or matrix,
but also as a string describing a transformation from the affinity matrix (e.g. "x -> -log(x)").
}
\examples{
library(ConScapeR)
library(terra)

# Launch Julia
ConScapeR_setup(Sys.getenv("JULIA_PATH"))

# Create a SpatRaster from a file for the landscape permeability or affinities and habitat suitability
aff <- terra::rast(system.file("data/affinities_2000.asc", package="ConScapeR"))
hab <- terra::rast(system.file("data/suitability_2000.asc", package="ConScapeR"))

# Create a ConScape Grid
g <- ConScapeR::Grid(affinities=aff, sources=hab, targets=hab, costs="x -> -log(x)")
}
