% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{mat2rast}
\alias{mat2rast}
\title{Convert a matrix to raster}
\usage{
mat2rast(mat, rast)
}
\arguments{
\item{mat}{\verb{[matrix]} matrix to be converted}

\item{rast}{\verb{[SpatRaster]} template raster}
}
\description{
Convert a matrix to raster
}
